requirements.txt
pandas
ta
python-dotenv
